Add runlog and rename file
